Counter game from HackerRank webpage.
It's a game where we strat from number 5 and if it's a power of 2 player divide it by 2 and pass it to the next player but if it's not a power of 2 we're searching for the closest power of 2(but lower than n) and substract. Player wins if he/she reaches the n = 1 in his/her round.